done few commands. Need workaround with reloading gui from commands

+++ resources/lib/plexbmc.py
@@ -3929,41 +3929,6 @@ def displayServers( url ):
 
     xbmcplugin.endOfDirectory(pluginhandle,cacheToDisc=settings.get_setting('kodicache'))
 
-def switch_user():
-    # Get list of users
-    user_list = plex_network.get_plex_home_users()
-    # zero means we are not plexHome'd up
-    if user_list is None or len(user_list) == 1:
-        printDebug("No users listed or only one user, plexHome not enabled")
-        return False
-
-    printDebug("found %s users: %s" % (len(user_list), user_list.keys()))
-
-    # Get rid of currently logged in user.
-    user_list.pop(plex_network.get_myplex_user(), None)
-
-    select_screen = xbmcgui.Dialog()
-    result = select_screen.select('Switch User', user_list.keys())
-    if result == -1:
-        printDebug("Dialog cancelled")
-        return False
-
-    printDebug("user [%s] selected" % user_list.keys()[result])
-    user = user_list[user_list.keys()[result]]
-
-    pin = None
-    if user['protected'] == '1':
-        printDebug("Protected user [%s], requesting password" % user['title'])
-        pin = select_screen.input("Enter PIN", type=xbmcgui.INPUT_NUMERIC, option=xbmcgui.ALPHANUM_HIDE_INPUT)
-
-    success, msg = plex_network.switch_plex_home_user(user['id'], pin)
-
-    if not success:
-        xbmcgui.Dialog().ok("Switch Failed",msg)
-        return False
-
-    return True
-
 # So this is where we really start the addon
 printDebug = PrintDebug("PleXBMC")
 
@@ -4034,11 +3999,11 @@ def start_plexbmc():
     # todo pass parameters to constructor
     command = COMMANDS.get(command_name)()
 
-    '''if command_name == "cacherefresh":
+    # region Done commands
+    """
+    if command_name == "cacherefresh":
         plex_network.delete_cache()
-        xbmc.executebuiltin("ReloadSkin()")'''
-    if command and isinstance(command, BaseCommand):
-        command.execute()
+        xbmc.executebuiltin("ReloadSkin()")
 
     # Open the add-on settings page, then refresh plugin
     elif command_name == "setting":
@@ -4066,6 +4031,11 @@ def start_plexbmc():
                 xbmc.executebuiltin("Container.Refresh")
         else:
             printDebug.info("Switch User Failed")
+    """
+    # endregion
+
+    if command and isinstance(command, BaseCommand):
+        command.execute()
 
     elif command_name == "signout":
         if not plex_network.is_admin():
